Split up type-formats.ts module
It's not uncommon for me to hit dependency cycles due to the static
initialization of types in prelude-types.ts. I may eventually switch to
lazy initialization, but before I do this is step 1 of an attempt to
reduce the blast radius of those dependency cycles. Upcoming commits
will factor things further.
