+++ src/language/semantics/type-system/type-formats/type.ts
@@ -1,0 +1,20 @@
+import type {
+  ApplicationType,
+  FunctionType,
+  IndexedAccessType,
+  IntrinsicApplicationType,
+  ObjectType,
+  TypeParameter,
+  UnionType,
+} from '../type-formats.js'
+import type { OpaqueType } from './opaque-type.js'
+
+export type Type =
+  | ApplicationType
+  | FunctionType
+  | IndexedAccessType
+  | IntrinsicApplicationType
+  | ObjectType
+  | OpaqueType
+  | TypeParameter
+  | UnionType
